+++ project/3rdparty/include/asio2/base/detail/util.hpp
@@ -116,8 +116,11 @@ namespace asio2::detail
 
 namespace asio2::detail
 {
+	struct tcp_tag { using tl_tag_type = tcp_tag; }; // transport layer
+	struct udp_tag { using tl_tag_type = udp_tag; }; // transport layer
+
 	struct ssl_stream_tag {};
-	struct ws_stream_tag {};
+	struct ws_stream_tag  {};
 }
 
 namespace asio2
@@ -745,6 +748,32 @@ namespace asio2
 			return detail::get_current_object<T>();
 		}
 	}
+}
+
+namespace asio2::detail
+{
+	struct wait_timer_op : public asio::coroutine
+	{
+		asio::steady_timer& timer_;
+
+		wait_timer_op(asio::steady_timer& timer) : timer_(timer)
+		{
+		}
+
+		template <typename Self>
+		void operator()(Self& self, error_code ec = {})
+		{
+			detail::ignore_unused(ec);
+
+			ASIO_CORO_REENTER(*this)
+			{
+				ASIO_CORO_YIELD
+					timer_.async_wait(std::move(self));
+
+				self.complete(get_last_error());
+			}
+		}
+	};
 }
 
 #endif // !__ASIO2_UTIL_HPP__