+++ project/3rdparty/include/asio2/http/http_execute.hpp
@@ -23,7 +23,7 @@
 #include <asio2/http/detail/http_make.hpp>
 #include <asio2/http/detail/http_traits.hpp>
 
-#include <asio2/component/socks/socks5_client.hpp>
+#include <asio2/component/socks/socks5_client_cp.hpp>
 
 namespace asio2::detail
 {
@@ -52,6 +52,11 @@ namespace asio2::detail
 			std::string_view h{ sk5->host() };
 			std::string_view p{ sk5->port() };
 
+			if (static_cast<int>(sk5->command()) == 0)
+			{
+				sk5->command(socks5::command::connect);
+			}
+
 			// Look up the domain name
 			resolver.async_resolve(h, p, [&, s5 = std::move(sk5)]
 			(const error_code& ec1, const asio::ip::tcp::resolver::results_type& endpoints) mutable
@@ -64,14 +69,13 @@ namespace asio2::detail
 				{
 					if (ec2) { set_last_error(ec2); return; }
 
-					detail::socks5_client_connect_op
-					{
-						ioc,
+					socks5_async_handshake
+					(
 						detail::to_string(std::forward<String  >(host)),
 						detail::to_string(std::forward<StrOrInt>(port)),
 						socket,
 						std::move(s5),
-						[&](error_code ecs5) mutable
+						[&](error_code ecs5, std::string, std::string) mutable
 						{
 							if (ecs5) { set_last_error(ecs5); return; }
 
@@ -90,7 +94,7 @@ namespace asio2::detail
 								});
 							});
 						}
-					};
+					);
 				});
 			});
 		}