+++ project/3rdparty/include/asio2/udp/udp_cast.hpp
@@ -53,12 +53,17 @@
 #include <asio2/udp/impl/udp_send_cp.hpp>
 #include <asio2/udp/impl/udp_send_op.hpp>
 
+#include <asio2/proxy/socks5_client.hpp>
+
 namespace asio2::detail
 {
-	struct template_args_udp_cast
+	struct template_args_udp_cast : public udp_tag
 	{
 		using socket_t = asio::ip::udp::socket;
 		using buffer_t = asio2::linear_buffer;
+
+		template<class derived_t>
+		using socks5_client_t = asio2::socks5_client_t<derived_t>;
 
 		static constexpr std::size_t function_storage_size = 88;
 		static constexpr std::size_t allocator_storage_size = 256;
@@ -82,6 +87,7 @@ namespace asio2::detail
 		, public condition_event_cp<derived_t, args_t>
 		, public udp_send_cp       <derived_t, args_t>
 		, public udp_send_op       <derived_t, args_t>
+		, public udp_tag
 	{
 		ASIO2_CLASS_FRIEND_DECLARE_BASE;
 		ASIO2_CLASS_FRIEND_DECLARE_UDP_BASE;