+++ project/3rdparty/include/sqlite/sqlite3.h
@@ -46,9 +46,9 @@ extern "C" {
 #undef SQLITE_VERSION_NUMBER
 #endif
 
-#define SQLITE_VERSION                              "3.42.0"
-#define SQLITE_VERSION_NUMBER                       3042000
-#define SQLITE_SOURCE_ID                            "2023-05-16 12:36:15 831d0fb2836b71c9bc51067c49fee4b8f18047814f2ff22d817d25195cf350b0"
+#define SQLITE_VERSION                              "3.43.0"
+#define SQLITE_VERSION_NUMBER                       3043000
+#define SQLITE_SOURCE_ID                            "2023-08-24 12:36:59 0f80b798b3f4b81a7bb4233c58294edd0f1156f36b6ecf5ab8e83631d468778c"
 
 SQLITE_API SQLITE_EXTERN const char sqlite3_version[];
 
@@ -165,6 +165,7 @@ SQLITE_API int sqlite3_exec(sqlite3 *, const char *sql, int (*callback)(void *, 
 #define SQLITE_IOERR_ROLLBACK_ATOMIC                (SQLITE_IOERR | (31<<8))
 #define SQLITE_IOERR_DATA                           (SQLITE_IOERR | (32<<8))
 #define SQLITE_IOERR_CORRUPTFS                      (SQLITE_IOERR | (33<<8))
+#define SQLITE_IOERR_IN_PAGE                        (SQLITE_IOERR | (34<<8))
 #define SQLITE_LOCKED_SHAREDCACHE                   (SQLITE_LOCKED |  (1<<8))
 #define SQLITE_LOCKED_VTAB                          (SQLITE_LOCKED |  (2<<8))
 #define SQLITE_BUSY_RECOVERY                        (SQLITE_BUSY   |  (1<<8))
@@ -604,6 +605,7 @@ SQLITE_API const char *sqlite3_normalized_sql(sqlite3_stmt *pStmt);
 
 SQLITE_API int sqlite3_stmt_readonly(sqlite3_stmt *pStmt);
 SQLITE_API int sqlite3_stmt_isexplain(sqlite3_stmt *pStmt);
+SQLITE_API int sqlite3_stmt_explain(sqlite3_stmt *pStmt, int eMode);
 
 SQLITE_API int sqlite3_stmt_busy(sqlite3_stmt *);
 
@@ -1027,7 +1029,8 @@ SQLITE_API int sqlite3_test_control(int op, ...);
 #define SQLITE_TESTCTRL_TRACEFLAGS              31
 #define SQLITE_TESTCTRL_TUNE                    32
 #define SQLITE_TESTCTRL_LOGEST                  33
-#define SQLITE_TESTCTRL_LAST                    33
+#define SQLITE_TESTCTRL_USELONGDOUBLE           34
+#define SQLITE_TESTCTRL_LAST                    34
 
 SQLITE_API int sqlite3_keyword_count(void);
 SQLITE_API int sqlite3_keyword_name(int,const char **, int *);
@@ -1507,9 +1510,9 @@ typedef struct fts5_api fts5_api;
 struct fts5_api {
     int iVersion;
 
-    int (*xCreateTokenizer)(fts5_api *pApi, const char *zName, void *pContext, fts5_tokenizer *pTokenizer, void (*xDestroy)(void *));
-    int (*xFindTokenizer)(fts5_api *pApi, const char *zName, void **ppContext, fts5_tokenizer *pTokenizer);
-    int (*xCreateFunction)(fts5_api *pApi, const char *zName, void *pContext, fts5_extension_function xFunction, void (*xDestroy)(void *));
+    int (*xCreateTokenizer)(fts5_api *pApi, const char *zName, void *pUserData, fts5_tokenizer *pTokenizer, void (*xDestroy)(void *));
+    int (*xFindTokenizer)(fts5_api *pApi, const char *zName, void **ppUserData, fts5_tokenizer *pTokenizer);
+    int (*xCreateFunction)(fts5_api *pApi, const char *zName, void *pUserData, fts5_extension_function xFunction, void (*xDestroy)(void *));
 };
 
 #ifdef __cplusplus